requirementx.txt
numpy
geopandas
dask
xarray
rioxarray
tqdm